event entity

+++ src/main/groovy/pl/jug/torun/xenia/rest/dto/EventResponse.groovy
@@ -4,7 +4,7 @@ package pl.jug.torun.xenia.rest.dto
  * Created by mephi_000 on 06.09.14.
  */
 class EventResponse {
-    int id
+    long id
     String title
     String startDate
     String endDate
